clean: Simplify and improve automatic generation of release notes DOCS-206

+++ Makefile
@@ -41,17 +41,15 @@ helm_dep_up:
 update_dependencies: setup_helm_repos helm_dep_up update_internals_versions update_docs_versions
 
 .PHONY: get_release_notes
-get_release_notes: create_version_file
+get_release_notes:
+	# Obtain current and previous chart versions from the chart release branches
+	$(eval VERSION_NEW=$(shell git branch --remote --sort version:refname | grep -Eo "release-[0-9]+\.[0-9]+\.[0-9]+$$" | tail -n 1 | cut -d - -f 2))
+	$(eval VERSION_OLD=$(shell git branch --remote --sort version:refname | grep -Eo "release-[0-9]+\.[0-9]+\.[0-9]+$$" | tail -n 2 | head -n 1 | cut -d - -f 2))
+	if [ -z "${VERSION_NEW}" ] || [ -z "${VERSION_OLD}" ]; then echo "Can't find the current or previous chart version from the release branches."; exit 1; fi
+	# Clone codacy/codacy-tools-release-notes
+	if [ -d codacy-tools-release-notes ]; then rm -rf codacy-tools-release-notes; fi
+	git clone git@github.com:codacy/codacy-tools-release-notes.git
 	# Fetch updated codacy/chart tags
 	git fetch --all --tags --force
-	# Cleanup codacy-tools-release-notes if it exists
-	if [ -d codacy-tools-release-notes ]; then rm -rf codacy-tools-release-notes; fi
-	# Clone codacy/codacy-tools-release-notes and generate the changelog and release notes
-	git clone git@github.com:codacy/codacy-tools-release-notes.git
-	# Install dependencies
-	pip3 install -r codacy-tools-release-notes/requirements.pip --user
-	# Generate changelogs and release notes
-	codacy-tools-release-notes/getChangelogs.sh \
-		$(shell cat .version | grep -Eoh "^([0-9]+\.[0-9]+\.[0-9]+)") \
-		${CURDIR}/codacy/requirements.yaml \
-		${CURDIR}/codacy/requirements.lock
+	# Generate changelog and release notes
+	codacy-tools-release-notes/run.sh ${VERSION_NEW} ${VERSION_OLD}
